Refactor imports and disable sgDesigner scoring
Reorganized and cleaned up imports across modules to reduce circular dependencies and clarify module responsibilities (crisprhawk.py, scoring.py, search_guides.py). Moved and re-ordered imports (annotation, candidate_guides, converter, scoring_envs, reports, etc.), removed unused imports (Bitset, subprocess), and adjusted haplotype/region/guide imports. Temporarily commented out the sgDesigner scoring invocation in scoring_guides to avoid invoking the external environment until env handling is updated. Minor haplotype import adjustments in search_guides.py.
