+++ src/crisprhawk/crisprhawk.py
@@ -6,29 +6,29 @@ configuration management. It provides high-level functions that orchestrate the
 various steps of the CRISPR-HAWK analysis pipeline.
 """
 
-from .config_utils import ScoringEnvs
+from .annotation import annotate_guides
+from .candidate_guides import candidate_guides_analysis
+from .converter import convert_gnomad_vcf
 from .crisprhawk_argparse import (
     CrisprHawkSearchInputArgs,
     CrisprHawkConverterInputArgs,
     CrisprHawkPrepareDataInputArgs,
 )
-from .region_constructor import construct_regions
-from .haplotypes import reconstruct_haplotypes
-from .haplotype import Haplotype
-from .region import Region
-from .utils import print_verbosity, VERBOSITYLVL
-from .search_guides import search
-from .annotation import annotate_guides
-from .scoring import scoring_guides
-from .search_offtargets import offtargets_search
-from .encoder import encode
-from .reports import report_guides
-from .graphical_reports import compute_graphical_reports
-from .converter import convert_gnomad_vcf
 from .crisprme_data import prepare_data_crisprme
-from .candidate_guides import candidate_guides_analysis
+from .encoder import encode
+from .graphical_reports import compute_graphical_reports
 from .guide import Guide
+from .haplotype import Haplotype
+from .haplotypes import reconstruct_haplotypes
 from .pam import PAM
+from .region import Region
+from .region_constructor import construct_regions
+from .reports import report_guides
+from .scoring import scoring_guides
+from .scoring_envs import ScoringEnvs
+from .search_guides import search
+from .search_offtargets import offtargets_search
+from .utils import print_verbosity, VERBOSITYLVL
 
 from typing import Dict, List
 from time import time
